+++ Assets/Tests/InputSystem/Plugins/UITests.cs
@@ -543,9 +543,9 @@ internal class UITests : InputTestFixture
             // Create actions.
             var map = new InputActionMap("map");
             asset.AddActionMap(map);
-            var moveAction = map.AddAction("move");
-            var submitAction = map.AddAction("submit");
-            var cancelAction = map.AddAction("cancel");
+            var moveAction = map.AddAction("move", type: InputActionType.PassThrough);
+            var submitAction = map.AddAction("submit", type: InputActionType.PassThrough);
+            var cancelAction = map.AddAction("cancel", type: InputActionType.PassThrough);
 
             // Create bindings.
             moveAction.AddBinding(gamepads[i].leftStick);
@@ -661,9 +661,9 @@ internal class UITests : InputTestFixture
         // Create actions.
         var map = new InputActionMap("map");
         asset.AddActionMap(map);
-        var moveAction = map.AddAction("move");
-        var submitAction = map.AddAction("submit");
-        var cancelAction = map.AddAction("cancel");
+        var moveAction = map.AddAction("move", type: InputActionType.PassThrough);
+        var submitAction = map.AddAction("submit", type: InputActionType.PassThrough);
+        var cancelAction = map.AddAction("cancel", type: InputActionType.PassThrough);
 
         // Create bindings.
         moveAction.AddBinding(gamepad.leftStick);
@@ -764,8 +764,8 @@ internal class UITests : InputTestFixture
                     {
                         ""name"" : ""UI"",
                         ""actions"" : [
-                            { ""name"" : ""Navigate"" },
-                            { ""name"" : ""Point"" }
+                            { ""name"" : ""Navigate"", ""type"" : ""PassThrough"" },
+                            { ""name"" : ""Point"", ""type"" : ""PassThrough"" }
                         ]
                     }
                 ]
@@ -868,6 +868,131 @@ internal class UITests : InputTestFixture
             Press(mouse.leftButton);
             Profiler.EndSample();
         }, Is.Not.AllocatingGCMemory());
+    }
+
+    // https://forum.unity.com/threads/feature-request-option-to-disable-deselect-in-ui-input-module.761531
+    [UnityTest]
+    [Category("UI")]
+    public IEnumerator UI_CanPreventAutomaticDeselectionOfGameObjects()
+    {
+        var mouse = InputSystem.AddDevice<Mouse>();
+
+        var actions = ScriptableObject.CreateInstance<InputActionAsset>();
+        var uiActions = actions.AddActionMap("UI");
+        var pointAction = uiActions.AddAction("Point", type: InputActionType.PassThrough, binding: "<Mouse>/position");
+        var clickAction = uiActions.AddAction("Click", type: InputActionType.PassThrough, binding: "<Mouse>/leftButton");
+
+        actions.Enable();
+
+        var testObjects = CreateScene(0, 200);
+
+        testObjects.uiModule.actionsAsset = actions;
+        testObjects.uiModule.point = InputActionReference.Create(pointAction);
+        testObjects.uiModule.leftClick = InputActionReference.Create(clickAction);
+
+        // Deselect behavior should be on by default as this corresponds to the behavior before
+        // we introduced the switch that allows toggling the behavior off.
+        Assert.That(testObjects.uiModule.deselectOnBackgroundClick, Is.True);
+
+        // Give canvas a chance to set itself up.
+        yield return null;
+
+        // Click on left GO and make sure it gets selected.
+        Set(mouse.position, new Vector2(10, 10));
+        Press(mouse.leftButton);
+        yield return null;
+        Release(mouse.leftButton);
+
+        Assert.That(testObjects.eventSystem.currentSelectedGameObject, Is.SameAs(testObjects.leftGameObject));
+
+        // Click outside of GOs and make sure the selection gets cleared.
+        Set(mouse.position, new Vector2(50, 250));
+        Press(mouse.leftButton);
+        yield return null;
+        Release(mouse.leftButton);
+
+        Assert.That(testObjects.eventSystem.currentSelectedGameObject, Is.Null);
+
+        testObjects.uiModule.deselectOnBackgroundClick = false;
+
+        // Click on left GO and make sure it gets selected.
+        Set(mouse.position, new Vector2(10, 10));
+        Press(mouse.leftButton);
+        yield return null;
+        Release(mouse.leftButton);
+
+        Assert.That(testObjects.eventSystem.currentSelectedGameObject, Is.SameAs(testObjects.leftGameObject));
+
+        // Click outside of GOs and make sure our selection does NOT get cleared.
+        Set(mouse.position, new Vector2(50, 250));
+        Press(mouse.leftButton);
+        yield return null;
+        Release(mouse.leftButton);
+
+        Assert.That(testObjects.eventSystem.currentSelectedGameObject, Is.SameAs(testObjects.leftGameObject));
+    }
+
+    ////REVIEW: While `deselectOnBackgroundClick` does solve the problem of breaking keyboard and gamepad navigation, the question
+    ////        IMO is whether navigation should even be affected that way by not having a current selection. Seems to me that the
+    ////        the system should remember the last selected object and start up navigation from there when nothing is selected.
+    ////        However, given EventSystem.lastSelectedGameObject is no longer supported (why???), it seems like this would require
+    ////        some larger changes.
+    [UnityTest]
+    [Category("UI")]
+    [Ignore("TODO")]
+    public IEnumerator TODO_UI_CanStartNavigationWhenNothingIsSelected()
+    {
+        var mouse = InputSystem.AddDevice<Mouse>();
+        var gamepad = InputSystem.AddDevice<Gamepad>();
+
+        var actions = ScriptableObject.CreateInstance<InputActionAsset>();
+        var uiActions = actions.AddActionMap("UI");
+        var pointAction = uiActions.AddAction("Point", type: InputActionType.PassThrough, binding: "<Mouse>/position");
+        var clickAction = uiActions.AddAction("Click", type: InputActionType.PassThrough, binding: "<Mouse>/leftButton");
+        var navigateAction = uiActions.AddAction("Navigate", type: InputActionType.PassThrough, binding: "<Gamepad>/dpad");
+
+        actions.Enable();
+
+        var testObjects = CreateScene(0, 200);
+
+        testObjects.uiModule.actionsAsset = actions;
+        testObjects.uiModule.point = InputActionReference.Create(pointAction);
+        testObjects.uiModule.leftClick = InputActionReference.Create(clickAction);
+        testObjects.uiModule.move = InputActionReference.Create(navigateAction);
+
+        // Give canvas a chance to set itself up.
+        yield return null;
+
+        // Select left GO.
+        Set(mouse.position, new Vector2(10, 10));
+        Press(mouse.leftButton);
+        yield return null;
+        Release(mouse.leftButton);
+
+        Assert.That(testObjects.eventSystem.currentSelectedGameObject, Is.SameAs(testObjects.leftGameObject));
+
+        // Click on background and make sure we deselect.
+        Set(mouse.position, new Vector2(50, 250));
+        Press(mouse.leftButton);
+        yield return null;
+        Release(mouse.leftButton);
+
+        Assert.That(testObjects.eventSystem.currentSelectedGameObject, Is.Null);
+
+        // Now perform a navigate-right action. Given we have no current selection, this should
+        // cause the right GO to be selected based on the fact that the left GO was selected last.
+        Press(gamepad.dpad.right);
+        yield return null;
+
+        Assert.That(testObjects.eventSystem.currentSelectedGameObject, Is.SameAs(testObjects.rightGameObject));
+
+        // Just to make extra sure, navigate left and make sure that results in the expected selection
+        // change over to the left GO.
+        Release(gamepad.dpad.right);
+        Press(gamepad.dpad.left);
+        yield return null;
+
+        Assert.That(testObjects.eventSystem.currentSelectedGameObject, Is.SameAs(testObjects.leftGameObject));
     }
 
 // The tracked device tests fail with NullReferenceException in the windows editor on yamato. I cannot reproduce this locally, so will disable them on windows for now.
@@ -1369,6 +1494,14 @@ internal class UITests : InputTestFixture
         public void InvokeUpdate()
         {
             Update();
+        }
+
+        protected override void OnApplicationFocus(bool hasFocus)
+        {
+            // Sync our focus state to that of the test runtime rather than to the Unity test runner (where
+            // debugging may still focus and thus alter the test run).
+            hasFocus = ((InputTestRuntime)InputRuntime.s_Instance).hasFocus;
+            base.OnApplicationFocus(hasFocus);
         }
     }
 }